Write local backups directly to disk

+++ app/DatabaseBackup.php
@@ -5,6 +5,7 @@ namespace App;
 use Illuminate\Filesystem\FilesystemAdapter;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
+use League\Flysystem\Adapter\Local;
 use League\Flysystem\Filesystem;
 use League\Flysystem\WebDAV\WebDAVAdapter;
 use Spatie\DbDumper\DbDumper;
@@ -13,6 +14,19 @@ class DatabaseBackup
 {
     public function __construct(private DbDumper $dumper, private FilesystemAdapter $filesystem)
     {
+    }
+
+    /**
+     * Returns a boolean whether the filesystem is a Local driver.
+     *
+     * @return boolean
+     */
+    private function isLocal(): bool
+    {
+        /** @var Filesystem $driver */
+        $driver = $this->filesystem->getDriver();
+
+        return $driver->getAdapter() instanceof Local;
     }
 
     /**
@@ -38,11 +52,23 @@ class DatabaseBackup
     {
         $filename = Str::slug($database) . '-' . now()->format('Y-m-d-H-i-s') . '.sql.gz';
 
+        $isLocal = $this->isLocal();
+
         try {
-            $this->dumper->setDbName($database)->dumpToFile(Storage::path($filename));
-            $this->filesystem->put($filename, $this->isWebdav() ? Storage::get($filename) : Storage::readStream($filename));
+            $this->dumper->setDbName($database)->dumpToFile(
+                $isLocal ? $this->filesystem->path($filename) : Storage::path($filename)
+            );
+
+            if (!$isLocal) {
+                $this->filesystem->put(
+                    $filename,
+                    $this->isWebdav() ? Storage::get($filename) : Storage::readStream($filename)
+                );
+            }
         } finally {
-            Storage::delete($filename);
+            if (!$isLocal) {
+                Storage::delete($filename);
+            }
         }
 
         return $filename;
